Move layouts back to src
Without this change managing PYTHONPATH and imports was tricky.
E.g. it was hard to properly support both bin/bundle and building
individual layouts.

Move all layouts to src so that now only one directory needs to be on
PYTHONPATH and also all imports throughout the project are relative to
this one directory.